test: extract agent definition white-box suite
